# Рефакторинг Вынес все команды в отдельные домены
