+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "greenlet>=3.2.3",
     "httpx>=0.28.1",
     "kafka-python>=2.2.15",
+    "logger>=1.4",
     "pydantic>=2.11.7",
     "pydantic-settings>=2.10.1",
     "sqlalchemy>=2.0.41",
