+++ src/tg_bot/domains/russian_roulette/__init__.py
@@ -1,5 +1,5 @@
 from aiogram import Router
 
-russian_roulette_router = Router()
+russian_roulette_router = Router(name="russian_roulette_router")
 
 import src.tg_bot.domains.russian_roulette.handlers
