+++ src/tg_bot/domains/russian_roulette/services.py
@@ -2,12 +2,13 @@ from dataclasses import dataclass
 
 from aiogram import types
 from dishka import FromDishka
-from dishka.integrations.aiogram import inject
 
 from src.app.users.schemas import UsersCreateSchema
 from src.tg_bot.domains.russian_roulette.repository import RussianRouletteRepository
 from src.tg_bot.domains.russian_roulette.schemas import BadPhraseMessage
-from src.tg_bot.domains.user_management.services import UserBotService, extract_user_data
+from src.tg_bot.domains.user_management.services import (
+    extract_user_data,
+)
 
 
 @dataclass
