+++ src/tg_bot/domains/start_domain/__init__.py
@@ -1,0 +1,5 @@
+from aiogram import Router
+
+start_router = Router(name="start_router")
+
+from src.tg_bot.domains.start_domain import handlers
