+++ src/tg_bot/domains/user_management/__init__.py
@@ -1,5 +1,5 @@
 from aiogram import Router
 
-user_router = Router()
+user_router = Router(name="user_router")
 
 from src.tg_bot.domains.user_management import handlers
